feat: rename @zetapush/core to @zetapush/client
affects: @zetapush/cli, @zetapush/client, @zetapush/create, @zetapush/example,
@zetapush/integration, @zetapush/worker

BREAKING CHANGE:
Rename @zetapush/core to @zetapush/client

ISSUES CLOSED: #79
