+++ packages/client/package.json
@@ -1,8 +1,8 @@
 {
-  "name": "@zetapush/core",
+  "name": "@zetapush/client",
   "version": "0.29.0",
   "description": "Zeta Push Next",
-  "browser": "dist/zetapush.js",
+  "browser": "dist/zetapush-client.js",
   "main": "lib/index.js",
   "module": "es/index.js",
   "jsnext:main": "es/index.js",
@@ -20,8 +20,8 @@
     "clean": "rimraf lib dist es .deploy .esdoc",
     "build:commonjs": "cross-env BABEL_ENV=commonjs babel src --out-dir lib",
     "build:es": "cross-env BABEL_ENV=es babel src --out-dir es",
-    "build:umd": "cross-env BABEL_ENV=umd NODE_ENV=development rollup -c -i src/index.js -o dist/zetapush.js",
-    "build:umd:min": "cross-env BABEL_ENV=umd NODE_ENV=production rollup -c -i src/index.js -o dist/zetapush.min.js",
+    "build:umd": "cross-env BABEL_ENV=umd NODE_ENV=development rollup -c -i src/index.js -o dist/zetapush-client.js",
+    "build:umd:min": "cross-env BABEL_ENV=umd NODE_ENV=production rollup -c -i src/index.js -o dist/zetapush-client.min.js",
     "build": "npm run build:commonjs && npm run build:es && npm run build:umd && npm run build:umd:min",
     "prepublish": "npm run clean && npm run build",
     "test:ci": "karma start karma.conf.js"
