+++ packages/client/src/authentication/handshake.js
