+++ packages/client/src/client/smart.js
