+++ packages/client/src/utils/prototype.js
