+++ packages/client/src/utils/session-persistence.js
