+++ packages/client/test/.eslintrc
