+++ packages/client/test/client-error.spec.js
