+++ packages/client/test/lifecycle-events.spec.js
