+++ packages/client/test/prerequisites.spec.js
