+++ packages/client/test/weak-client.spec.js
