+++ packages/create/template/javascript/front/index.html
@@ -58,7 +58,7 @@
   <nav>
     <button class="js-Hello">hello</button>
   </nav>
-  <script src="https://unpkg.com/@zetapush/core"></script>
+  <script src="https://unpkg.com/@zetapush/client"></script>
   <script src="https://unpkg.com/@zetapush/platform"></script>
   <script src="./index.js"></script>
 </body>
