+++ packages/example/front/index.html
@@ -60,7 +60,7 @@
     <button class="js-Add">add</button>
     <button class="js-List">list</button>
   </nav>
-  <script src="https://unpkg.com/@zetapush/core"></script>
+  <script src="https://unpkg.com/@zetapush/client"></script>
   <script src="https://unpkg.com/@zetapush/platform"></script>
   <script src="./index.js"></script>
 </body>
